admission: re-home KVAdmissionController interface/impl
In a future PR we'll spruce up the "handle" type and something other
than the opaque interface{} it is today. We're doing it to be able to
integrate with elastic CPU tokens we're introducing for export requests
-- we want to use the "handle" given to us by the KVAdmissionController,
thread it into the surrounding context, rely on it having already
plumbed into storage.MVCCExportToSST and use the "handle" to check
whether we're over our allotted CPU slice before returning. All this
feels context more appropriate to sit within the admission package, and
this commit paves the way for it.

Release note: None
