+++ pkg/util/admission/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "grant_coordinator.go",
         "granter.go",
         "io_load_listener.go",
+        "kv_admission_controller.go",
         "kv_slot_adjuster.go",
         "sql_cpu_overload_indicator.go",
         "store_token_estimation.go",
@@ -22,10 +23,12 @@ go_library(
         "//pkg/roachpb",
         "//pkg/settings",
         "//pkg/settings/cluster",
+        "//pkg/util",
         "//pkg/util/admission/admissionpb",
         "//pkg/util/humanizeutil",
         "//pkg/util/log",
         "//pkg/util/metric",
+        "//pkg/util/stop",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
